+++ dio_cache_interceptor/lib/src/util/response_extension.dart
@@ -7,18 +7,19 @@ extension ResponseExtension on Response {
   /// https://tools.ietf.org/html/rfc7232#section-4.1
   void updateCacheHeaders(Response<dynamic>? response) {
     if (response == null) return;
-    _updateNonNullHeader(cacheControlHeader, response);
-    _updateNonNullHeader(dateHeader, response);
-    _updateNonNullHeader(etagHeader, response);
-    _updateNonNullHeader(lastModifiedHeader, response);
-    _updateNonNullHeader(expiresHeader, response);
-    _updateNonNullHeader(contentLocationHeader, response);
-    _updateNonNullHeader(varyHeader, response);
-  }
 
-  void _updateNonNullHeader(String headerKey, Response<dynamic> response) {
-    final values = response.headers[headerKey];
-    if (values != null) headers.map[headerKey] = values;
+    void updateNonNullHeader(String headerKey) {
+      final values = response.headers[headerKey];
+      if (values != null) headers.map[headerKey] = values;
+    }
+
+    updateNonNullHeader(cacheControlHeader);
+    updateNonNullHeader(dateHeader);
+    updateNonNullHeader(etagHeader);
+    updateNonNullHeader(lastModifiedHeader);
+    updateNonNullHeader(expiresHeader);
+    updateNonNullHeader(contentLocationHeader);
+    updateNonNullHeader(varyHeader);
   }
 
   /// Checks if disposition of the response is attachment